# [mypyc] Use native calls in singledispatch dispatch functions