This adds support for using native calls to registered implementations when generating dispatch functions for singledispatch functions, instead of always using non-native calls, which can significantly speed up singledispatch (see benchmark results below). We don't use native calls when the registered implementation is a decorated function, as using a native call would ignore any changes the decorator made to the function.

### Benchmark Results

These are the results from running the sum_tree_singledispatch both with this change and on the current master (7808e8265cd4cb29dc7f05ff58ab945ab3f2d20b).

#### Current master
```
interpreted: 0.694200s (avg of 5 iterations; stdev 1%)
compiled:    0.661199s (avg of 5 iterations; stdev 1.3%)

compiled is 1.050x faster
```
#### This PR
```
interpreted: 0.701401s (avg of 5 iterations; stdev 1.4%)
compiled:    0.603199s (avg of 5 iterations; stdev 1.7%)

compiled is 1.163x faster
```
## Test Plan

I added an irbuild test to make sure we are actually using native calls when there are no non-register decorators.